+++ code/src/cljs/sixsq/nuvla/ui/credentials/views.cljs
@@ -431,6 +431,39 @@
           (@tr [:nuvlabox-modal-more-info])]]))))
 
 
+(defn credential-openstack
+  []
+  (let [tr             (subscribe [::i18n-subs/tr])
+        is-new?        (subscribe [::subs/is-new?])
+        credential     (subscribe [::subs/credential])
+        validate-form? (subscribe [::subs/validate-form?])
+        on-change      (fn [name-kw value]
+                         (dispatch [::events/update-credential name-kw value])
+                         (dispatch [::events/validate-credential-form ::spec/openstack-credential]))]
+    (fn []
+      (let [editable? (utils-general/editable? @credential @is-new?)
+            {:keys [name description openstack-username openstack-password]} @credential]
+        [:<>
+         [ui/Table style/definition
+          [ui/TableBody
+           [uix/TableRowField (@tr [:name]), :editable? editable?, :required? true,
+            :default-value name, :spec ::spec/name, :validate-form? @validate-form?,
+            :on-change (partial on-change :name)]
+           [uix/TableRowField (@tr [:description]), :editable? editable?, :required? true,
+            :default-value description, :spec ::spec/description, :validate-form? @validate-form?,
+            :on-change (partial on-change :description)]
+           [uix/TableRowField "username", :placeholder "OpenStack username", :editable? editable?, :required? true,
+            :default-value openstack-username, :spec ::spec/openstack-username, :validate-form? @validate-form?,
+            :on-change (partial on-change :openstack-username)]
+           [uix/TableRowField "password", :placeholder "OpenStack password", :editable? editable?, :required? true,
+            :default-value openstack-password, :spec ::spec/openstack-password, :validate-form? @validate-form?,
+            :on-change (partial on-change :openstack-password)]]]
+         [:div {:style {:color "grey" :font-style "oblique"}} (@tr [:credential-cloud-follow-link])]
+         [:a {:href   "https://docs.openstack.org/api-ref/identity/v3/index.html#authentication-and-token-management"
+              :target "_blank"}
+          (@tr [:nuvlabox-modal-more-info])]]))))
+
+
 (defn save-callback
   [form-validation-spec]
   (dispatch-sync [::events/set-validate-form? true])
@@ -453,7 +486,10 @@
     :modal-content   credential-azure}
    "infrastructure-service-google"
    {:validation-spec ::spec/google-credential
-    :modal-content   credential-google}})
+    :modal-content   credential-google}
+   "infrastructure-service-openstack"
+   {:validation-spec ::spec/openstack-credential
+    :modal-content   credential-openstack}})
 
 
 (def infrastructure-service-csp-subtypes
@@ -555,6 +591,7 @@
     "infrastructure-service-google" {:tab-index (:cloud-services tab-indices), :icon "cloud", :name "Google Compute"}
     "infrastructure-service-amazonec2" {:tab-index (:cloud-services tab-indices), :icon "cloud", :name "AWS EC2"}
     "infrastructure-service-exoscale" {:tab-index (:cloud-services tab-indices), :icon "cloud", :name "Exoscale"}
+    "infrastructure-service-openstack" {:tab-index (:cloud-services tab-indices), :icon "cloud", :name "OpenStack"}
     "infrastructure-service-vpn" {:tab-index (:access-services tab-indices), :icon "key", :name "VPN"}
     "gpg-key" {:tab-index (:access-services tab-indices), :icon "key", :name "GPG keys"}
     "ssh-key" {:tab-index (:access-services tab-indices), :icon "key", :name "SSH keys"}
@@ -800,7 +837,20 @@
           [ui/CardContent {:text-align :center}
            [ui/Header "Cloud Google"]
            [ui/Image {:src   "/ui/images/gce.png"
-                      :style {:max-width 130}}]]]]]])))
+                      :style {:max-width 130}}]]]
+
+
+         [ui/Card
+          {:on-click #(do
+                        (dispatch [::events/set-validate-form? false])
+                        (dispatch [::events/form-valid])
+                        (dispatch [::events/close-add-credential-modal])
+                        (dispatch [::events/open-credential-modal
+                                   {:subtype "infrastructure-service-openstack"} true]))}
+          [ui/CardContent {:text-align :center}
+           [ui/Header "Cloud OpenStack"]
+           [ui/Image {:src   "/ui/images/openstack.png"
+                      :style {:max-width 220}}]]]]]])))
 
 
 (defn generated-credential-modal